requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.109.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
oauthlib==3.2.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlmodel==0.0.14
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1

dateparser~=1.2.0
spacy~=3.7.4
tqdm~=4.66.2
httpx~=0.28.1
pillow~=11.0.0